+++ src/components/Contact.css
@@ -1,4 +1,4 @@
-.booking-form {
+.contact-form {
   max-width: 600px;
   background: white;
   padding: 30px;
@@ -6,6 +6,6 @@
   box-shadow: 0px 0px 20px rgba(0,0,0,0.2);
 }
 
-.booking-form label {
+.contact-form label {
   font-weight: 600;
 }
