+++ src/components/Navbar.jsx
@@ -15,7 +15,7 @@ export default function Navbar() {
         <ul className="navbar-nav gap-3">
           <li className="nav-item"><a className="nav-link" href="#hero">Inicio</a></li>
           <li className="nav-item"><a className="nav-link" href="#rooms">Habitaciones</a></li>
-          <li className="nav-item"><a className="nav-link" href="#booking">Reservas</a></li>
+          <li className="nav-item"><a className="nav-link" href="#booking">Contáctenos</a></li>
           <li className="nav-item"><a className="nav-link" href="#contact">Contacto</a></li>
         </ul>
       </div>
